Separate fields tests into separate modules (date/datetime/complexdatetime) relates to #1983
