+++ tests/fields/fields.py
@@ -10,11 +10,6 @@ import sys
 from nose.plugins.skip import SkipTest
 import six
 
-try:
-    import dateutil
-except ImportError:
-    dateutil = None
-
 from decimal import Decimal
 
 from bson import Binary, DBRef, ObjectId, SON
@@ -30,54 +25,8 @@ from mongoengine.base import (BaseDict, BaseField, EmbeddedDocumentList,
 
 from tests.utils import MongoDBTestCase
 
-__all__ = ("FieldTest", "EmbeddedDocumentListFieldTestCase")
-
 
 class FieldTest(MongoDBTestCase):
-
-    def test_datetime_from_empty_string(self):
-        """
-        Ensure an exception is raised when trying to
-        cast an empty string to datetime.
-        """
-        class MyDoc(Document):
-            dt = DateTimeField()
-
-        md = MyDoc(dt='')
-        self.assertRaises(ValidationError, md.save)
-
-    def test_date_from_empty_string(self):
-        """
-        Ensure an exception is raised when trying to
-        cast an empty string to datetime.
-        """
-        class MyDoc(Document):
-            dt = DateField()
-
-        md = MyDoc(dt='')
-        self.assertRaises(ValidationError, md.save)
-
-    def test_datetime_from_whitespace_string(self):
-        """
-        Ensure an exception is raised when trying to
-        cast a whitespace-only string to datetime.
-        """
-        class MyDoc(Document):
-            dt = DateTimeField()
-
-        md = MyDoc(dt='   ')
-        self.assertRaises(ValidationError, md.save)
-
-    def test_date_from_whitespace_string(self):
-        """
-        Ensure an exception is raised when trying to
-        cast a whitespace-only string to datetime.
-        """
-        class MyDoc(Document):
-            dt = DateField()
-
-        md = MyDoc(dt='   ')
-        self.assertRaises(ValidationError, md.save)
 
     def test_default_values_nothing_set(self):
         """Ensure that default field values are used when creating
@@ -694,273 +643,6 @@ class FieldTest(MongoDBTestCase):
         for api_key in invalid:
             person.api_key = api_key
             self.assertRaises(ValidationError, person.validate)
-
-    def test_datetime_validation(self):
-        """Ensure that invalid values cannot be assigned to datetime
-        fields.
-        """
-        class LogEntry(Document):
-            time = DateTimeField()
-
-        log = LogEntry()
-        log.time = datetime.datetime.now()
-        log.validate()
-
-        log.time = datetime.date.today()
-        log.validate()
-
-        log.time = datetime.datetime.now().isoformat(' ')
-        log.validate()
-
-        if dateutil:
-            log.time = datetime.datetime.now().isoformat('T')
-            log.validate()
-
-        log.time = -1
-        self.assertRaises(ValidationError, log.validate)
-        log.time = 'ABC'
-        self.assertRaises(ValidationError, log.validate)
-
-    def test_date_validation(self):
-        """Ensure that invalid values cannot be assigned to datetime
-        fields.
-        """
-        class LogEntry(Document):
-            time = DateField()
-
-        log = LogEntry()
-        log.time = datetime.datetime.now()
-        log.validate()
-
-        log.time = datetime.date.today()
-        log.validate()
-
-        log.time = datetime.datetime.now().isoformat(' ')
-        log.validate()
-
-        if dateutil:
-            log.time = datetime.datetime.now().isoformat('T')
-            log.validate()
-
-        log.time = -1
-        self.assertRaises(ValidationError, log.validate)
-        log.time = 'ABC'
-        self.assertRaises(ValidationError, log.validate)
-
-    def test_datetime_tz_aware_mark_as_changed(self):
-        from mongoengine import connection
-
-        # Reset the connections
-        connection._connection_settings = {}
-        connection._connections = {}
-        connection._dbs = {}
-
-        connect(db='mongoenginetest', tz_aware=True)
-
-        class LogEntry(Document):
-            time = DateTimeField()
-
-        LogEntry.drop_collection()
-
-        LogEntry(time=datetime.datetime(2013, 1, 1, 0, 0, 0)).save()
-
-        log = LogEntry.objects.first()
-        log.time = datetime.datetime(2013, 1, 1, 0, 0, 0)
-        self.assertEqual(['time'], log._changed_fields)
-
-    def test_datetime(self):
-        """Tests showing pymongo datetime fields handling of microseconds.
-        Microseconds are rounded to the nearest millisecond and pre UTC
-        handling is wonky.
-
-        See: http://api.mongodb.org/python/current/api/bson/son.html#dt
-        """
-        class LogEntry(Document):
-            date = DateTimeField()
-
-        LogEntry.drop_collection()
-
-        # Test can save dates
-        log = LogEntry()
-        log.date = datetime.date.today()
-        log.save()
-        log.reload()
-        self.assertEqual(log.date.date(), datetime.date.today())
-
-        # Post UTC - microseconds are rounded (down) nearest millisecond and
-        # dropped
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 999)
-        d2 = datetime.datetime(1970, 1, 1, 0, 0, 1)
-        log = LogEntry()
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertNotEqual(log.date, d1)
-        self.assertEqual(log.date, d2)
-
-        # Post UTC - microseconds are rounded (down) nearest millisecond
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 9999)
-        d2 = datetime.datetime(1970, 1, 1, 0, 0, 1, 9000)
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertNotEqual(log.date, d1)
-        self.assertEqual(log.date, d2)
-
-        if not six.PY3:
-            # Pre UTC dates microseconds below 1000 are dropped
-            # This does not seem to be true in PY3
-            d1 = datetime.datetime(1969, 12, 31, 23, 59, 59, 999)
-            d2 = datetime.datetime(1969, 12, 31, 23, 59, 59)
-            log.date = d1
-            log.save()
-            log.reload()
-            self.assertNotEqual(log.date, d1)
-            self.assertEqual(log.date, d2)
-
-    def test_date(self):
-        """Tests showing pymongo date fields
-
-        See: http://api.mongodb.org/python/current/api/bson/son.html#dt
-        """
-        class LogEntry(Document):
-            date = DateField()
-
-        LogEntry.drop_collection()
-
-        # Test can save dates
-        log = LogEntry()
-        log.date = datetime.date.today()
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, datetime.date.today())
-
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 999)
-        d2 = datetime.datetime(1970, 1, 1, 0, 0, 1)
-        log = LogEntry()
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, d1.date())
-        self.assertEqual(log.date, d2.date())
-
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 9999)
-        d2 = datetime.datetime(1970, 1, 1, 0, 0, 1, 9000)
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, d1.date())
-        self.assertEqual(log.date, d2.date())
-
-        if not six.PY3:
-            # Pre UTC dates microseconds below 1000 are dropped
-            # This does not seem to be true in PY3
-            d1 = datetime.datetime(1969, 12, 31, 23, 59, 59, 999)
-            d2 = datetime.datetime(1969, 12, 31, 23, 59, 59)
-            log.date = d1
-            log.save()
-            log.reload()
-            self.assertEqual(log.date, d1.date())
-            self.assertEqual(log.date, d2.date())
-
-    def test_datetime_usage(self):
-        """Tests for regular datetime fields"""
-        class LogEntry(Document):
-            date = DateTimeField()
-
-        LogEntry.drop_collection()
-
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1)
-        log = LogEntry()
-        log.date = d1
-        log.validate()
-        log.save()
-
-        for query in (d1, d1.isoformat(' ')):
-            log1 = LogEntry.objects.get(date=query)
-            self.assertEqual(log, log1)
-
-        if dateutil:
-            log1 = LogEntry.objects.get(date=d1.isoformat('T'))
-            self.assertEqual(log, log1)
-
-        # create additional 19 log entries for a total of 20
-        for i in range(1971, 1990):
-            d = datetime.datetime(i, 1, 1, 0, 0, 1)
-            LogEntry(date=d).save()
-
-        self.assertEqual(LogEntry.objects.count(), 20)
-
-        # Test ordering
-        logs = LogEntry.objects.order_by("date")
-        i = 0
-        while i < 19:
-            self.assertTrue(logs[i].date <= logs[i + 1].date)
-            i += 1
-
-        logs = LogEntry.objects.order_by("-date")
-        i = 0
-        while i < 19:
-            self.assertTrue(logs[i].date >= logs[i + 1].date)
-            i += 1
-
-        # Test searching
-        logs = LogEntry.objects.filter(date__gte=datetime.datetime(1980, 1, 1))
-        self.assertEqual(logs.count(), 10)
-
-        logs = LogEntry.objects.filter(date__lte=datetime.datetime(1980, 1, 1))
-        self.assertEqual(logs.count(), 10)
-
-        logs = LogEntry.objects.filter(
-            date__lte=datetime.datetime(1980, 1, 1),
-            date__gte=datetime.datetime(1975, 1, 1),
-        )
-        self.assertEqual(logs.count(), 5)
-
-    def test_date_usage(self):
-        """Tests for regular datetime fields"""
-        class LogEntry(Document):
-            date = DateField()
-
-        LogEntry.drop_collection()
-
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1)
-        log = LogEntry()
-        log.date = d1
-        log.validate()
-        log.save()
-
-        for query in (d1, d1.isoformat(' ')):
-            log1 = LogEntry.objects.get(date=query)
-            self.assertEqual(log, log1)
-
-        if dateutil:
-            log1 = LogEntry.objects.get(date=d1.isoformat('T'))
-            self.assertEqual(log, log1)
-
-        # create additional 19 log entries for a total of 20
-        for i in range(1971, 1990):
-            d = datetime.datetime(i, 1, 1, 0, 0, 1)
-            LogEntry(date=d).save()
-
-        self.assertEqual(LogEntry.objects.count(), 20)
-
-        # Test ordering
-        logs = LogEntry.objects.order_by("date")
-        i = 0
-        while i < 19:
-            self.assertTrue(logs[i].date <= logs[i + 1].date)
-            i += 1
-
-        logs = LogEntry.objects.order_by("-date")
-        i = 0
-        while i < 19:
-            self.assertTrue(logs[i].date >= logs[i + 1].date)
-            i += 1
-
-        # Test searching
-        logs = LogEntry.objects.filter(date__gte=datetime.datetime(1980, 1, 1))
-        self.assertEqual(logs.count(), 10)
 
     def test_list_validation(self):
         """Ensure that a list field only accepts lists with valid elements."""
@@ -1699,7 +1381,7 @@ class FieldTest(MongoDBTestCase):
         post.save()
 
         post = BlogPost()
-        post.info = {'title' : 'dollar_sign', 'details' : {'te$t' : 'test'} }
+        post.info = {'title': 'dollar_sign', 'details': {'te$t': 'test'}}
         post.save()
 
         post = BlogPost()
@@ -1718,7 +1400,7 @@ class FieldTest(MongoDBTestCase):
 
         post = BlogPost.objects.filter(info__title__exact='dollar_sign').first()
         self.assertIn('te$t', post['info']['details'])
-         
+
         # Confirm handles non strings or non existing keys
         self.assertEqual(
             BlogPost.objects.filter(info__details__test__exact=5).count(), 0)
@@ -5398,181 +5080,6 @@ class GenericLazyReferenceFieldTest(MongoDBTestCase):
         occ.in_embedded.direct = animal1_ref
         occ.in_embedded.in_list = [animal1_ref, animal2_ref]
         check_fields_type(occ)
-
-
-class ComplexDateTimeFieldTest(MongoDBTestCase):
-    def test_complexdatetime_storage(self):
-        """Tests for complex datetime fields - which can handle
-        microseconds without rounding.
-        """
-        class LogEntry(Document):
-            date = ComplexDateTimeField()
-            date_with_dots = ComplexDateTimeField(separator='.')
-
-        LogEntry.drop_collection()
-
-        # Post UTC - microseconds are rounded (down) nearest millisecond and
-        # dropped - with default datetimefields
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 999)
-        log = LogEntry()
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, d1)
-
-        # Post UTC - microseconds are rounded (down) nearest millisecond - with
-        # default datetimefields
-        d1 = datetime.datetime(1970, 1, 1, 0, 0, 1, 9999)
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, d1)
-
-        # Pre UTC dates microseconds below 1000 are dropped - with default
-        # datetimefields
-        d1 = datetime.datetime(1969, 12, 31, 23, 59, 59, 999)
-        log.date = d1
-        log.save()
-        log.reload()
-        self.assertEqual(log.date, d1)
-
-        # Pre UTC microseconds above 1000 is wonky - with default datetimefields
-        # log.date has an invalid microsecond value so I can't construct
-        # a date to compare.
-        for i in range(1001, 3113, 33):
-            d1 = datetime.datetime(1969, 12, 31, 23, 59, 59, i)
-            log.date = d1
-            log.save()
-            log.reload()
-            self.assertEqual(log.date, d1)
-            log1 = LogEntry.objects.get(date=d1)
-            self.assertEqual(log, log1)
-
-        # Test string padding
-        microsecond = map(int, [math.pow(10, x) for x in range(6)])
-        mm = dd = hh = ii = ss = [1, 10]
-
-        for values in itertools.product([2014], mm, dd, hh, ii, ss, microsecond):
-            stored = LogEntry(date=datetime.datetime(*values)).to_mongo()['date']
-            self.assertTrue(re.match('^\d{4},\d{2},\d{2},\d{2},\d{2},\d{2},\d{6}$', stored) is not None)
-
-        # Test separator
-        stored = LogEntry(date_with_dots=datetime.datetime(2014, 1, 1)).to_mongo()['date_with_dots']
-        self.assertTrue(re.match('^\d{4}.\d{2}.\d{2}.\d{2}.\d{2}.\d{2}.\d{6}$', stored) is not None)
-
-    def test_complexdatetime_usage(self):
-        """Tests for complex datetime fields - which can handle
-        microseconds without rounding.
-        """
-        class LogEntry(Document):
-            date = ComplexDateTimeField()
-
-        LogEntry.drop_collection()
-
-        d1 = datetime.datetime(1950, 1, 1, 0, 0, 1, 999)
-        log = LogEntry()
-        log.date = d1
-        log.save()
-
-        log1 = LogEntry.objects.get(date=d1)
-        self.assertEqual(log, log1)
-
-        # create extra 59 log entries for a total of 60
-        for i in range(1951, 2010):
-            d = datetime.datetime(i, 1, 1, 0, 0, 1, 999)
-            LogEntry(date=d).save()
-
-        self.assertEqual(LogEntry.objects.count(), 60)
-
-        # Test ordering
-        logs = LogEntry.objects.order_by("date")
-        i = 0
-        while i < 59:
-            self.assertTrue(logs[i].date <= logs[i + 1].date)
-            i += 1
-
-        logs = LogEntry.objects.order_by("-date")
-        i = 0
-        while i < 59:
-            self.assertTrue(logs[i].date >= logs[i + 1].date)
-            i += 1
-
-        # Test searching
-        logs = LogEntry.objects.filter(date__gte=datetime.datetime(1980, 1, 1))
-        self.assertEqual(logs.count(), 30)
-
-        logs = LogEntry.objects.filter(date__lte=datetime.datetime(1980, 1, 1))
-        self.assertEqual(logs.count(), 30)
-
-        logs = LogEntry.objects.filter(
-            date__lte=datetime.datetime(2011, 1, 1),
-            date__gte=datetime.datetime(2000, 1, 1),
-        )
-        self.assertEqual(logs.count(), 10)
-
-        LogEntry.drop_collection()
-
-        # Test microsecond-level ordering/filtering
-        for microsecond in (99, 999, 9999, 10000):
-            LogEntry(
-                date=datetime.datetime(2015, 1, 1, 0, 0, 0, microsecond)
-            ).save()
-
-        logs = list(LogEntry.objects.order_by('date'))
-        for next_idx, log in enumerate(logs[:-1], start=1):
-            next_log = logs[next_idx]
-            self.assertTrue(log.date < next_log.date)
-
-        logs = list(LogEntry.objects.order_by('-date'))
-        for next_idx, log in enumerate(logs[:-1], start=1):
-            next_log = logs[next_idx]
-            self.assertTrue(log.date > next_log.date)
-
-        logs = LogEntry.objects.filter(
-            date__lte=datetime.datetime(2015, 1, 1, 0, 0, 0, 10000))
-        self.assertEqual(logs.count(), 4)
-
-    def test_no_default_value(self):
-        class Log(Document):
-            timestamp = ComplexDateTimeField()
-
-        Log.drop_collection()
-
-        log = Log()
-        self.assertIsNone(log.timestamp)
-        log.save()
-
-        fetched_log = Log.objects.with_id(log.id)
-        self.assertIsNone(fetched_log.timestamp)
-
-    def test_default_static_value(self):
-        NOW = datetime.datetime.utcnow()
-        class Log(Document):
-            timestamp = ComplexDateTimeField(default=NOW)
-
-        Log.drop_collection()
-
-        log = Log()
-        self.assertEqual(log.timestamp, NOW)
-        log.save()
-
-        fetched_log = Log.objects.with_id(log.id)
-        self.assertEqual(fetched_log.timestamp, NOW)
-
-    def test_default_callable(self):
-        NOW = datetime.datetime.utcnow()
-
-        class Log(Document):
-            timestamp = ComplexDateTimeField(default=datetime.datetime.utcnow)
-
-        Log.drop_collection()
-
-        log = Log()
-        self.assertGreaterEqual(log.timestamp, NOW)
-        log.save()
-
-        fetched_log = Log.objects.with_id(log.id)
-        self.assertGreaterEqual(fetched_log.timestamp, NOW)
 
 
 if __name__ == '__main__':
